requirements.txt
Django==5.2.7
psycopg2-binary==2.9.11
Pillow==12.0.0
django-cors-headers==4.9.0
django-extensions==4.1
python-decouple==3.8
